Update Makefile (#2147)
* Problem: it's difficult to run/test/compile docs

Solution: we already have a nice Makefile, but it's outdated. The idea
is to revamp it and make it easy to use.

* Problem: Makefile is not documented

Solution: Add basic instructions on how to use it in the main README.md

* Problem: can't connect to localhost:9984

Solution: make docker expose 9984 by default, so everyone can connect to
localhost:9984.

* Problem: make clean is TMI

Solution: add `@` to remove commands so we don't output all the details
about the cleaning

* Problem: make clean is too shy

Solution: print a message saying that cleaning went well.